split gallery and product
