+++ backend/app/api/routes/gallery.py
@@ -1,0 +1,600 @@
+from typing import Any
+import random
+
+from fastapi import (
+    APIRouter,
+    HTTPException,
+    Query,
+    BackgroundTasks,
+    Request
+)
+from app.core.deps import supabase
+from app.core.logging import get_logger
+from app.core.utils import slugify, generate_sku
+from app.models.generic import Message, ImageBulkDelete
+from app.models.product import (
+    ProductImageMetadata,
+    ImagesBulkUpdate,
+    ProductImageBulkUrls
+)
+from app.services.meilisearch import get_or_create_index, ensure_index_ready
+from app.prisma_client import prisma as db
+from app.core.config import settings
+from prisma.errors import UniqueViolationError
+from app.services.product import index_images, reindex_image, sync_index_image
+from app.services.redis import cache_response
+from meilisearch.errors import MeilisearchApiError
+from app.services.recently_viewed import RecentlyViewedService
+from app.services.websocket import manager
+
+logger = get_logger(__name__)
+
+
+async def process_bulk_delete_images(images: list):
+    """
+    Process bulk deletion of gallery images.
+    """
+    await manager.broadcast_to_all(
+        data={"status": "processing"},
+        message_type="product_bulk_delete"
+    )
+
+    try:
+        deleted_count = 0
+        failed_deletions = []
+
+        for index, image in enumerate(images):
+            try:
+                if image.image and "/storage/v1/object/public/product-images/" in image.image:
+                    try:
+                        file_path = image.image.split(
+                            "/storage/v1/object/public/product-images/")[1]
+                        supabase.storage.from_(
+                            "product-images").remove([file_path])
+                    except Exception as storage_error:
+                        logger.warning(
+                            f"Failed to delete from storage: {storage_error}")
+
+                await db.productimage.delete(where={"id": image.id})
+                deleted_count += 1
+
+                logger.info(
+                    f"Successfully deleted image {index + 1}/{len(images)}")
+
+            except Exception as e:
+                logger.error(f"Failed to delete image {index + 1}: {str(e)}")
+                failed_deletions.append({
+                    "image_id": image.id,
+                    "error": str(e)
+                })
+
+        for image in images:
+            await sync_index_image(image_id=str(image.id), product=image.product)
+
+        await manager.broadcast_to_all(
+            data={"status": "completed"},
+            message_type="product_bulk_delete"
+        )
+
+        logger.info(
+            f"Bulk delete completed successfully. {deleted_count} deleted, {len(failed_deletions)} failed.")
+
+    except Exception as e:
+        logger.error(f"Error processing bulk delete: {str(e)}")
+        await manager.broadcast_to_all(
+            data={"status": "failed"},
+            message_type="product_bulk_delete"
+        )
+
+
+def build_variant_data(payload) -> dict[str, Any]:
+    data = {}
+    if payload.size is not None:
+        data["size"] = payload.size
+    if payload.color is not None:
+        data["color"] = payload.color
+    if payload.measurement is not None:
+        data["measurement"] = payload.measurement
+    if payload.inventory is not None:
+        data["inventory"] = payload.inventory
+        data["status"] = "IN_STOCK" if payload.inventory > 0 else "OUT_OF_STOCK"
+    if payload.price is not None:
+        data["price"] = payload.price
+    if payload.old_price is not None:
+        data["old_price"] = payload.old_price
+    return data
+
+
+def build_relation_data(category_ids=None, collection_ids=None) -> dict[str, Any]:
+    data: dict[str, Any] = {}
+    if category_ids:
+        data["categories"] = {"connect": [{"id": id} for id in category_ids]}
+    if collection_ids:
+        data["collections"] = {"connect": [{"id": id}
+                                           for id in collection_ids]}
+    return data
+
+
+async def handle_bulk_update_products(payload: ImagesBulkUpdate, images):
+    await manager.broadcast_to_all(
+        data={"status": "processing"},
+        message_type="product_bulk_update"
+    )
+    async with db.tx() as tx:
+        for index, image in enumerate(images):
+            try:
+                if image.product_id is None:
+                    name = f"{random.choice(['Classic', 'Premium', 'Superior', 'Deluxe', 'Luxury'])} {random.randint(100, 999)}"
+                    product_data: dict[str, Any] = {
+                        "name": name,
+                        "slug": slugify(name),
+                        "sku": generate_sku(),
+                        "active": payload.data.active,
+                        **build_relation_data(payload.data.category_ids, payload.data.collection_ids),
+                    }
+                    product = await tx.product.create(data=product_data)
+
+                    await tx.productimage.update(
+                        where={"id": image.id},
+                        data={"product": {"connect": {"id": product.id}}},
+                    )
+
+                    variant_data = {"sku": generate_sku(
+                    ), "product_id": product.id, "price": payload.data.price or 1, "inventory": payload.data.inventory or 0, **build_variant_data(payload.data)}
+                    await tx.productvariant.create(data=variant_data)
+
+                else:
+                    relation_updates: dict[str, Any] = {}
+                    if payload.data.category_ids:
+                        relation_updates["categories"] = {
+                            "set": [{"id": id} for id in payload.data.category_ids]}
+                    if payload.data.collection_ids:
+                        relation_updates["collections"] = {
+                            "set": [{"id": id} for id in payload.data.collection_ids]}
+
+                    if relation_updates:
+                        await tx.product.update(where={"id": image.product_id}, data=relation_updates)
+
+                    variant_data = build_variant_data(payload.data)
+                    if variant_data and image.product.variants:
+                        await tx.productvariant.update(
+                            where={"id": image.product.variants[0].id},
+                            data=variant_data,
+                        )
+            except Exception as e:
+                logger.error(f"Error processing image {image.id}: {e}")
+                continue
+
+    for image in images:
+        await reindex_image(image_id=image.id)
+
+    await manager.broadcast_to_all(
+        data={"status": "completed"},
+        message_type="product_bulk_update"
+    )
+
+router = APIRouter()
+
+
+# @router.get("/gallery3")
+# @cache_response("products:gallery3")
+# async def image_gallery(
+#     request: Request,
+#     skip: int = Query(default=0, ge=0),
+#     limit: int = Query(default=10, le=100),
+# ):
+#     """
+#     Lightweight gallery endpoint.
+#     Only returns product images and basic product info.
+#     Relations (categories, collections, variants, shared_collections)
+#     should be fetched on-demand in a detail endpoint.
+#     """
+#     try:
+#         # ✅ raw SQL for optimized SELECT
+#         query = """
+#         SELECT pi.id, pi.image, pi.product_id,
+#                 p.name AS product_name,
+#                 p.slug AS product_slug
+#         FROM "product_images" pi
+#         LEFT JOIN "products" p ON p.id = pi.product_id
+#         WHERE (COALESCE(pi."order", 0) = 0 OR pi.product_id IS NULL)
+#         ORDER BY pi.id DESC
+#         OFFSET $1 LIMIT $2
+#         """
+
+#         total_query = """
+#         SELECT COUNT(*)::int AS count
+#         FROM "product_images" pi
+#         LEFT JOIN "products" p ON p.id = pi.product_id
+#         WHERE (COALESCE(pi."order", 0) = 0 OR pi.product_id IS NULL)
+#         """
+
+#         async with database.pool.acquire() as conn:
+#             images = await conn.fetch(query, skip, limit)
+#             total = await conn.fetch(total_query)
+
+
+#         return {
+#             "images": [dict(record) for record in images],  # convert asyncpg.Record → dict
+#             "skip": skip,
+#             "limit": limit,
+#             "total_pages": ceil(total[0]["count"] / limit) if limit else 1,
+#             "total_count": total[0]["count"],
+#         }
+
+#     except Exception as e:
+#         logger.error(e)
+#         raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.get("/")
+@cache_response("products:gallery")
+async def index(
+    request: Request,
+    skip: int = Query(default=0, ge=0),
+    limit: int = Query(default=10, le=100),
+):
+    """
+    Retrieve products using Meilisearch, sorted by latest.
+    """
+
+    search_params = {
+        "limit": limit,
+        "offset": skip,
+        "sort": ["id:desc"],
+    }
+
+    index = get_or_create_index(settings.MEILI_IMAGES_INDEX)
+    try:
+        search_results = index.search(
+            "",
+            {
+                **search_params
+            }
+        )
+
+    except MeilisearchApiError as e:
+        error_code = getattr(e, "code", None)
+        if error_code in {"invalid_search_facets", "invalid_search_filter", "invalid_search_sort"}:
+            logger.warning(
+                f"Invalid filter detected, attempting to auto-configure filterable attributes: {e}")
+
+            ensure_index_ready(index)
+
+            search_results = index.search(
+                "",
+                {
+                    **search_params
+                }
+            )
+
+    except Exception as e:
+        logger.error(e)
+        raise HTTPException(
+            status_code=400,
+            detail=str(e)
+        )
+
+    total_count = search_results["estimatedTotalHits"]
+    total_pages = (total_count // limit) + (total_count % limit > 0)
+
+    return {
+        "images": search_results["hits"],
+        "skip": skip,
+        "limit": limit,
+        "total_count": total_count,
+        "total_pages": total_pages,
+    }
+
+@router.post("/reindex", response_model=Message)
+async def reindex_images(background_tasks: BackgroundTasks):
+    """
+    Re-index all images in the db to Meilisearch.
+    """
+    try:
+        background_tasks.add_task(index_images, invalidate_products=True)
+        return Message(message="Re-indexing task enqueued.")
+    except Exception as e:
+        logger.error(e)
+        raise HTTPException(
+            status_code=500,
+            detail=str(e),
+        )
+
+
+@router.delete("/{image_id}")
+async def delete_gallery_image(image_id: int) -> Message:
+    """
+    Delete a gallery image.
+    If the image is linked to a product, delete the entire product
+    """
+    image = await db.productimage.find_unique(where={"id": image_id})
+    if not image:
+        raise HTTPException(status_code=404, detail="Image not found")
+
+    if not image.product_id:
+        try:
+            file_path = image.image.split(
+                "/storage/v1/object/public/product-images/")[1]
+            supabase.storage.from_("product-images").remove([file_path])
+        except Exception as e:
+            logger.error(f"Error deleting image {image.image}: {e}")
+        await db.productimage.delete(where={"id": image_id})
+        await sync_index_image(image_id=image_id)
+
+        return Message(message="Image deleted successfully")
+
+    product = await db.product.find_unique(
+        where={"id": image.product_id},
+        include={
+            "shared_collections": True,
+        }
+    )
+
+    async with db.tx() as tx:
+        product_id = image.product_id
+
+        images = await tx.productimage.find_many(where={"product_id": product_id})
+        file_paths = [
+            img.image.split("/storage/v1/object/public/product-images/")[1]
+            for img in images
+            if "/storage/v1/object/public/product-images/" in img.image
+        ]
+
+        if file_paths:
+            supabase.storage.from_("product-images").remove(file_paths)
+
+        await tx.productimage.delete_many(where={"product_id": product_id})
+        await tx.review.delete_many(where={"product_id": product_id})
+        await tx.productvariant.delete_many(where={"product_id": product_id})
+        await tx.product.delete(where={"id": product_id})
+
+    service = RecentlyViewedService()
+    await service.remove_product_from_all(product_id=product_id)
+
+    await sync_index_image(image_id=image.id, product=product)
+
+    return Message(message="Product and all related data deleted successfully")
+
+
+@router.post("/bulk-upload")
+async def bulk_save_image_urls(payload: ProductImageBulkUrls):
+    """
+    Save many image URLs into the gallery.
+    """
+    try:
+        if not payload.urls:
+            raise HTTPException(status_code=400, detail="No images provided")
+
+        create_rows = []
+        for idx, url in enumerate(payload.urls):
+            create_rows.append({
+                "image": url,
+                "order": 0,
+            })
+
+        if not create_rows:
+            raise HTTPException(
+                status_code=400, detail="No valid images to save")
+
+        await db.productimage.create_many(data=create_rows)
+        await index_images()
+
+        return {"success": True, "count": len(create_rows)}
+    except HTTPException:
+        raise
+    except Exception as e:
+        logger.error(e)
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.post("/bulk-delete")
+async def bulk_delete_gallery_images(
+    payload: ImageBulkDelete,
+    background_tasks: BackgroundTasks,
+):
+    """
+    Bulk delete gallery images.
+    """
+    print(payload)
+    try:
+        if not payload.files:
+            raise HTTPException(
+                status_code=400, detail="No image IDs provided")
+
+        images = await db.productimage.find_many(
+            where={"id": {"in": payload.files}},
+            include={"product": True}
+        )
+
+        if not images:
+            raise HTTPException(status_code=404, detail="No images found")
+
+        background_tasks.add_task(process_bulk_delete_images, images=images)
+
+        return {
+            "success": True,
+            "message": f"Bulk delete job started. Processing {len(payload.files)} images in background."
+        }
+
+    except Exception as e:
+        logger.error(f"Error starting bulk delete job: {str(e)}")
+        raise HTTPException(status_code=400, detail=str(e))
+
+
+@router.post("/{image_id}/metadata")
+async def create_image_metadata(
+    image_id: int,
+    payload: ProductImageMetadata,
+    background_tasks: BackgroundTasks,
+):
+    """
+    Create a product, images and variants.
+    - Creates variants (if provided)
+    """
+    async with db.tx() as tx:
+        try:
+            slugified_name = slugify(payload.name)
+
+            data: dict[str, Any] = {
+                "name": payload.name,
+                "slug": slugified_name,
+                "sku": generate_sku(),
+                "description": payload.description,
+                "active": True,
+            }
+
+            if payload.category_ids:
+                data["categories"] = {"connect": [{"id": id}
+                                                  for id in payload.category_ids]}
+
+            if payload.collection_ids:
+                data["collections"] = {"connect": [{"id": id}
+                                                   for id in payload.collection_ids]}
+
+            product = await tx.product.create(data=data)
+
+            await tx.productimage.update(
+                where={"id": image_id},
+                data={
+                    "product": {"connect": {"id": product.id}}
+                }
+            )
+
+            if payload.variants:
+                for variant in payload.variants:
+                    try:
+                        await tx.productvariant.create(
+                            data={
+                                "sku": generate_sku(),
+                                "price": variant.price,
+                                "old_price": variant.old_price,
+                                "inventory": variant.inventory,
+                                "product_id": product.id,
+                                "status": variant.inventory > 0 and "IN_STOCK" or "OUT_OF_STOCK",
+                                "size": variant.size,
+                                "color": variant.color,
+                                "measurement": variant.measurement,
+                            }
+                        )
+                    except Exception as e:
+                        logger.error(e)
+                        raise HTTPException(
+                            status_code=400, detail=f"Failed to create variant: {str(e)}")
+
+            background_tasks.add_task(reindex_image, image_id=image_id)
+
+            return {"success": True}
+        except UniqueViolationError:
+            raise HTTPException(
+                status_code=400, detail="Product with this name already exists")
+        except HTTPException:
+            raise
+        except Exception as e:
+            logger.error(e)
+            raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.patch("/{image_id}/metadata")
+async def update_image_metadata(
+    image_id: int,
+    payload: ProductImageMetadata,
+    background_tasks: BackgroundTasks,
+):
+    """
+    - Updates product metadata
+    - Creates or updates variants (if provided)
+    """
+    async with db.tx() as tx:
+        try:
+            existing_image = await db.productimage.find_unique(where={"id": image_id})
+            if not existing_image:
+                raise HTTPException(status_code=404, detail="Image not found")
+
+            update_data = {}
+
+            if payload.name is not None:
+                update_data["name"] = payload.name
+                update_data["slug"] = slugify(payload.name)
+
+            if payload.description is not None:
+                update_data["description"] = payload.description
+
+            if payload.category_ids is not None:
+                category_ids = [{"id": id} for id in payload.category_ids]
+                update_data["categories"] = {"set": category_ids}
+
+            if payload.collection_ids is not None:
+                collection_ids = [{"id": id} for id in payload.collection_ids]
+                update_data["collections"] = {"set": collection_ids}
+            if payload.active is not None:
+                update_data["active"] = payload.active
+
+            await tx.product.update(
+                where={"id": existing_image.product_id},
+                data=update_data,
+            )
+
+            for variant in payload.variants:
+                try:
+                    if variant.id:
+                        await tx.productvariant.update(
+                            where={
+                                'id': variant.id
+                            },
+                            data={
+                                'price': variant.price,
+                                'old_price': variant.old_price,
+                                'inventory': variant.inventory,
+                                'status': variant.inventory > 0 and "IN_STOCK" or "OUT_OF_STOCK",
+                                'size': variant.size,
+                                'color': variant.color,
+                                'measurement': variant.measurement,
+                            }
+                        )
+                    else:
+                        await tx.productvariant.create(
+                            data={
+                                "product_id": existing_image.product_id,
+                                'sku': generate_sku(),
+                                'price': variant.price,
+                                'old_price': variant.old_price,
+                                'inventory': variant.inventory,
+                                'status': variant.inventory > 0 and "IN_STOCK" or "OUT_OF_STOCK",
+                                'size': variant.size,
+                                'color': variant.color,
+                                'measurement': variant.measurement,
+                            }
+                        )
+                except Exception as e:
+                    logger.error(e)
+                    raise HTTPException(
+                        status_code=400, detail=f"Failed to create variant: {str(e)}")
+
+            background_tasks.add_task(
+                reindex_image, image_id=existing_image.id)
+
+            return {"success": True}
+        except UniqueViolationError:
+            raise HTTPException(
+                status_code=400, detail="Product with this name already exists")
+        except HTTPException:
+            raise
+        except Exception as e:
+            logger.error(e)
+            raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.patch("/bulk-update")
+async def bulk_update_products(payload: ImagesBulkUpdate, background_tasks: BackgroundTasks):
+    images = await db.productimage.find_many(
+        where={"id": {"in": payload.image_ids}},
+        include={"product": {"include": {"variants": True}}},
+    )
+
+    if not images or len(images) != len(payload.image_ids):
+        raise HTTPException(status_code=404, detail="Some images not found")
+
+    background_tasks.add_task(handle_bulk_update_products, payload, images)
+
+    return {"message": "Products update started"}
