+++ frontend/lib/hooks/useProduct.ts
@@ -24,8 +24,7 @@ export const useProductSearch = (params: SearchParams) => {
     });
 };
 
-export const 
-useProductInfiniteSearch = (params: SearchParams) => {
+export const useProductInfiniteSearch = (params: SearchParams) => {
     return useInfiniteQuery({
         queryKey: ["products", "search", "infinite", JSON.stringify(params)],
         queryFn: async ({ pageParam = 0 }) =>
@@ -76,47 +75,6 @@ export const useSimilarProducts = (productId: number, num: number = 16) => {
             return response.json();
         },
         enabled: !!productId,
-    });
-};
-
-export const useImageGalleryInfinite = (pageSize: number = 24) => {
-    return useInfiniteQuery({
-        queryKey: ["products", "gallery"],
-        queryFn: async ({ pageParam = 0 }) =>
-            await api.get<{ images: SearchImageItem[]; skip: number; limit: number; total_count: number; total_pages: number }>("/product/gallery", {
-                params: { skip: pageParam, limit: pageSize },
-            }),
-        getNextPageParam: (lastPage) => {
-            const nextSkip = lastPage.skip + lastPage.limit;
-            const hasMore = nextSkip < lastPage.total_count;
-
-            return hasMore ? nextSkip : undefined;
-        },
-        initialPageParam: 0,
-    });
-};
-
-export const useCreateImageMetadata = () => {
-    return useMutation({
-        mutationFn: async ({ imageId, input }: { imageId: number; input: any }) => await api.post<Message>(`/product/${imageId}/metadata`, input),
-        onSuccess: () => {
-            toast.success("Image metadata created");
-        },
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to create image metadata");
-        },
-    });
-};
-
-export const useUpdateImageMetadata = () => {
-    return useMutation({
-        mutationFn: async ({ imageId, input }: { imageId: number; input: any }) => await api.patch<Message>(`/product/${imageId}/metadata`, input),
-        onSuccess: () => {
-            toast.success("Image metadata updated");
-        },
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to update image metadata");
-        },
     });
 };
 
@@ -285,49 +243,6 @@ export const useDeleteImages = () => {
         },
         onError: (error: any) => {
             toast.error(error.message || "Failed to delete image");
-        },
-    });
-};
-
-export const useDeleteGalleryImage = () => {
-    return useMutation({
-        mutationFn: async ({ id }: { id: number }) => await api.delete<Message>(`/product/gallery/${id}`),
-        onSuccess: () => {
-            toast.success("Image deleted successfully");
-        },
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to delete image");
-        },
-    });
-};
-
-export const useBulkDeleteGalleryImages = () => {
-    return useMutation({
-        mutationFn: async ({ imageIds }: { imageIds: number[] }) => await api.post<Message>(`/product/images/bulk-delete`, { files: imageIds }),
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to start bulk delete");
-        },
-    });
-};
-
-export const useBulkProductUpdate = () => {
-    return useMutation({
-        mutationFn: async ({ imageIds, input }: { imageIds: number[]; input: any }) =>
-            await api.patch<Message>("/product/gallery/bulk-update", { image_ids: imageIds, data: input }),
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to start bulk metadata");
-        },
-    });
-};
-
-export const useBulkSaveImageUrls = () => {
-    return useMutation({
-        mutationFn: async ({ urls }: { urls: string[] }) => await api.post<Message>("/product/images/bulk-save-urls", { urls }),
-        onSuccess: () => {
-            toast.success("Bulk image URLs saved successfully");
-        },
-        onError: (error: any) => {
-            toast.error(error.message || "Failed to save bulk image URLs");
         },
     });
 };
